chore: biome + prettier → oxlint + oxfmt (#37)
Replacing both linting and formatting with Oxc tools. They are quick,
and having both tools under a single umbrella is super nice.

## Linting

Replace Biome with
[Oxlint](https://oxc.rs/docs/guide/usage/linter.html).

## Formatting

Replace Prettier with
[Oxfmt](https://oxc.rs/docs/guide/usage/formatter.html).

This is mostly compatible with Prettier. We can not do file-type based
overrides yet, so we are unable to have 4-space Typescript and 2-space
YAML/package.json/etc.

We never used biome's formatter, because it didn't handle all of our
file types. Technically, oxfmt doesn't either, but it is kind enough to
hand off to prettier in those instances (ie. Markdown, HTML, etc).